+++ client/mds3/sampletable.js
@@ -357,7 +357,12 @@ async function createDiscoInSandbox(tk, block, sample, thisMutation) {
 	}
 	sandbox.header.text(headerTexts.join(''))
 	try {
-		;(await import('#plots/plot.disco.js')).default(tk.mds, tk.mds.label, sample, sandbox.body, block.genome)
+		;(await import('#plots/plot.disco.js')).default(tk.mds, tk.mds.label, sample, sandbox.body, block.genome, {
+			downloadImgName: headerTexts.join('') + ' Disco', // file name of svg downloaded from disco
+			label: {
+				prioritizeGeneLabelsByGeneSets: true // TODO control this at dataset
+			}
+		})
 	} catch (e) {
 		sandbox.body.append('div').text('Error: ' + (e.message || e))
 		if (e.stack) console.log(e.stack)