+++ client/plots/disco/interactions/DiscoInteractions.ts
@@ -3,8 +3,13 @@ export class DiscoInteractions {
 	downloadClickListener: (d: any) => void
 	geneClickListener: (gene: string, mnames: Array<string>) => void
 	prioritizeGenesCheckboxListener: (checked: boolean) => void
+	downloadImgName: string
 
 	constructor(app: any) {
+		// note! only call this constructor then app.state{} is created
+
+		this.downloadImgName = app.state.settings.downloadImgName || 'disco.plot'
+
 		this.cappingClickCallback = (d: any, t: any) => {
 			const tip = app.app.tip
 			tip.clear()
@@ -41,12 +46,13 @@ export class DiscoInteractions {
 
 			a.addEventListener(
 				'click',
-				function () {
+				() => {
+					// must use arrow function but not "function()", so this.downloadImgName is accessible
 					const serializer = new XMLSerializer()
 					const svg_blob = new Blob([serializer.serializeToString(svg)], {
 						type: 'image/svg+xml'
 					})
-					a.download = 'disco' + '.svg'
+					a.download = this.downloadImgName + '.svg'
 					a.href = URL.createObjectURL(svg_blob)
 					document.body.removeChild(a)
 				},