+++ client/plots/matrix.interactivity.js
@@ -186,7 +186,12 @@ export function setInteractivity(self) {
 						self.state.vocab.dslabel,
 						sample,
 						sandbox.body,
-						self.app.opts.genome
+						self.app.opts.genome,
+						{
+							label: {
+								prioritizeGeneLabelsByGeneSets: true // TODO control this at dataset-level
+							}
+						}
 					)
 				})
 		}