+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+Fixes:
+- In mass sampleview, expanding tree branch everytime no longer triggers rerendering of disco/ssgq plots
+- Disco: file name of downloaded svg preserves context including sample name and gene.
+- Disco: by default it now prioritize CGC genes in hg38.
+- Various scatterplot fixes: tooltip dot scale, divideBy bug
+- Fix the session recovery within an embedder portal
+
 
 ## 2.30.1
 
@@ -18,6 +25,7 @@ Features:
 
 Fixes:
 - Revert prev change that cause hic genome view to misalign
+
 
 ## 2.29.6